Add share gestures array in ShareAppDelegate. All gestures are shared across view controllers.

+++ TeamHyperFit/TeamHyperFit/Views/TTGraphView.m
@@ -153,7 +153,7 @@ const int kLineWidth        = 2;
         
         [self calGraphPointsWithData:self.data canvas:canvasRect pointsOut:graphPoints numberOfColum:self.numberOfColumn ];
         
-        NSLog(@"datasizebeforedrawline : %d",dataSize);
+        NSLog(@"datasizebeforedrawline : %ld",(long)dataSize);
         
         [self drawGraphLine:ctx points:graphPoints numberOfPoints:dataSize ];
     }
